Move publicTrades sync-schema to separate file
